feat(scheduler): add virtual machine scale set scheduler
Implemented a new scheduling capability for Azure
virtual machine scale set using the Terraform module.
This enhancement allows users to automate the start
and stop operations of their databases resources based
on the Azure Function timerTrigger.
The crontab expression is used to define
the desired time intervals for these actions.